feat(Module Component): Replace Renderer with a more generic Module component.
The Renderer was used for code splitting and generating a container div. This proved to be too high
level for a number of use cases. The new Module component replaces it. It gives more flexibility and
control to the user and should make the  code splitting point a bit more obvious.

BREAKING CHANGE: Replaced Renderer component.